Changed filename id format
The used id format (`date('Y-m-d H:i:s')`) for the generated files makes them impossible to sync with Windows clientes. I replaced  it with `date('YmdHis')` which makes more sense if we are using it as an id. I also renamed the variables from *date* to *id* because that is what that is, we are not using them as a timestamp but as an unique identifier for the filenames.

Fixes nextcloud/files_frommail#17

Signed-off-by: Alvaro Flaño <tikoflano@gmail.com>